Moved from tar-based form to phar-based.
SSPak now uses Phar as its storage format.  This makes for more elegant self-extracting phars,
and simplifies some of the code for reading/writing from the files.

Right now, the self-extractor doesn't have any knowledge of itself, and there isn't any
way of making a non-self-extracting file.  But it's a good start.